Thesis
Thesis discusses the study of the Measurement of lift force on aircraft wings or Airfoil in a low speed wind tunnel. The objective of thesis is to design the measurement of lifting force to get the most accurate value according to the simulation results by ANSYS program. And measure the Pressure Coefficient to determine the lifting force on the airfoil within a low speed wind tunnel It also studies and designs air filters for solve the streamline of air within the low speed wind tunnel. We are constructing an aircraft wing for testing in order to determine the lift force on the aircraft wing within the wind tunnel and use the pressure coefficient measuring instrument for test and calculate the lift force on the aircraft wing with wind tunnel Then, simulating the lift force on the aircraft wing with  ANSYS program and drawing in SolidWorks for low speed wind tunnel to determine the coefficient of pressure on the upper wing aircraft. The results of the study and test can be concluded that the lifting force on the Airfoil compared between the actual test and the simulation results with ANSYS program has an error for 17%. That Means that the measurement of lift force on the Airfoil within the wind tunnel is within the acceptable range. And in the future, can improve the wind tunnel structure, measuring tools, lifting force, or changing the shape of the Airfoil and can teaching in various subjects within the division of aircraft maintenance engineering technology.